refactor: 최종결제내역 컴포넌트 분리
