+++ app/purchase/[id]/page.tsx
@@ -1,12 +1,11 @@
 'use client';
 
-import { useEffect, useState } from 'react';
-import { RaffleItem } from '../../../lib/types/item';
-import useRaffleData from '../../../lib/hooks/useRaffleData';
+import { useState } from 'react';
 import Button from '../../../lib/common/Button';
 import PhoneNumber from '../../../components/PhoneNumber';
 import ShoppingAddressForm from '../../../components/payment/ShoppingAddressForm';
 import RaffleInfo from '../../../components/payment/RaffleInfo';
+import FinalPaymentSummary from '../../../components/payment/FinalPaymentSummary';
 
 export default function PurchasePage({
   params,
@@ -16,16 +15,7 @@ export default function PurchasePage({
   };
 }) {
   const { id } = params;
-  const [raffleItem, setRaffleItem] = useState<RaffleItem>({
-    ticketPrice: 0,
-    item: {
-      name: '',
-      description: '',
-      imageUrl: '',
-    },
-  });
   const [address, setAddress] = useState<string>('');
-  const { data, isLoading, isError, error } = useRaffleData();
 
   const handleSubmit = (event: React.FormEvent<HTMLFormElement>) => {
     event.preventDefault();
@@ -35,22 +25,6 @@ export default function PurchasePage({
   const handleAddressChange = (address: string) => {
     setAddress(address);
   };
-
-  useEffect(() => {
-    if (data && id) {
-      const foundItem = data.find((raffle: { id: number }) => raffle.id === parseInt(id));
-      if (foundItem) {
-        setRaffleItem(foundItem);
-      }
-    }
-  }, [data, id]);
-  if (isLoading) {
-    return <div>Loading...</div>;
-  }
-
-  if (isError || !raffleItem) {
-    return <div>Error: {error ? error.message : 'Item not found'}</div>;
-  }
 
   return (
     <main className="container mx-auto px-4 py-8">
@@ -70,9 +44,6 @@ export default function PurchasePage({
           </section>
 
           <section className="border p-4 rounded-md">
-            <div>
-              <h2 className="text-xl font-semibold mb-4">배송 정보</h2>
-            </div>
             <ShoppingAddressForm onAddressChange={handleAddressChange} />
             <div className="flex justify-center">
               <Button
@@ -89,11 +60,7 @@ export default function PurchasePage({
 
         <div className="space-y-4">
           <section className="border p-4 rounded-md">
-            <h2 className="text-xl font-semibold mb-4">최종 결제 내역</h2>
-            <div className="space-y-2">
-              <p>상품가격: {raffleItem.ticketPrice}원</p>
-              <p className="font-bold">총 결제 금액: {raffleItem.ticketPrice}원</p>
-            </div>
+            <FinalPaymentSummary id={id} />
           </section>
 
           <section className="border p-4 rounded-md">
