+++ app/purchase/layout.tsx
@@ -1,3 +1,3 @@
 export default function PuchaseLayout({ children }: { children: React.ReactNode }) {
-  return <div>{children}</div>;
+  return <>{children}</>;
 }
